refactor iceberg util
