+++ velox/functions/iceberg/CMakeLists.txt
@@ -12,4 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_subdirectory(util)
+velox_add_library(velox_functions_iceberg_util Murmur3Hash32.cpp)
+
+velox_link_libraries(velox_functions_iceberg_util velox_common_base
+                     velox_functions_util Folly::folly)
+
+if(${VELOX_BUILD_TESTING})
+  add_subdirectory(tests)
+endif()
